Use `@babel/core` for AST bundle processing and disallow HTML comments (#598)
* Use @babel/core for AST bundle processing and disallow HTML comments

* Disable Babel bundle optimisation

* Fix PR comments

* Merge post process functions

* Fix test name

* Use attachComment option

* Remove old config option
